Refactor: 숏츠탭 컴포넌트화 및 설명 주석
