+++ apps/service/src/constant/routes.ts
@@ -1,6 +1,7 @@
 export const ROUTES = {
   HOME: '/home',
   SHORTS: '/shorts',
+
   MYPAGE: '/mypage',
   SEARCH: (query?: string) => {
     if (!query?.trim()) return '/search';
